--- src/providers/ExtensionDetectionProvider/ExtensionDetectionRoot.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import type { ReactNode } from 'react';
-
-import { ExtensionDetectionProvider } from './ExtensionDetectionProvider';
-
-export function ExtensionDetectionRoot({ children }: { children: ReactNode }) {
-  return <ExtensionDetectionProvider>{children}</ExtensionDetectionProvider>;
-}
