--- src/providers/ExtensionDetectionProvider/extensionDetectionSingletonStore.ts
@@ -1,5 +1,0 @@
-'use client';
-
-import { createExtensionDetectionStore } from './store';
-
-export const extensionDetectionStore = createExtensionDetectionStore();
